feat: seed an unattributed address and an agent identity
