+++ seed/demo/PAGES.md
@@ -26,7 +26,8 @@ Status: `[x]` seeded + verified · `[~]` seeded, not yet verified · `[ ]` not s
 | Tool logs / traces                                             | CH `trace_summaries`; page enterprise-gated for 'demo' account type (README change 7)                                                                             | `[~]`  |
 | Insights (MCP & Tools)                                         | CH `trace_summaries`: unique per-surface trace ids + `gram.toolset.slug` (direct branch) + `gram.event.source=hook` rows (hook branch) + Skill hook rows          | `[~]`  |
 | Team page                                                      | PG `organization_user_relationships` + `users.workos_id` + role assignments (global_roles admin/member, skipped if absent)                                        | `[~]`  |
-| Employee enrollment                                            | PG memberships + `user_accounts` + `device_owners` + `device_agent_syncs`; roster served via impersonation carve-out                                              | `[~]`  |
+| Identities (roster + per-person pages) | PG memberships + `user_accounts`; CH telemetry keyed by email (agent-metrics view) and by `user.id` (raw logs) — one member-less address and one agent id so all three kinds render | `[~]`  |
+| Employee enrollment (detail only; index redirects to Identities)                                            | PG memberships + `user_accounts` + `device_owners` + `device_agent_syncs`; roster served via impersonation carve-out                                              | `[~]`  |
 | Org home (activity, facepiles, challenges)                     | PG `audit_logs` (22 rows: 12 general, 1 quarantine, 9 Killswitch lifecycle) + CH `authz_challenges` (13 rows incl. api_key bucket)                                | `[~]`  |
 | Audit logs                                                     | PG `audit_logs`: 9 canonical Killswitch actions (activate ×6, change, deactivate, expire), plus the existing general/quarantine history                           | `[~]`  |
 | Killswitch list                                                | PG `killswitch_prescriptions` ×6 current aggregates: Active ×3, Scheduled ×1, Lifted ×1, Expired ×1; selected/all scopes and one overlapping principal            | `[~]`  |
